Circle fixes: breakpoint actions structure
- Moved functions from `actions/breakpoints/index.js` over into
  `breakpoints.ts`, so that the index file is only re-exporting
